% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentMetadata.R
\name{DocumentMetadata}
\alias{DocumentMetadata}
\title{Document Metadata}
\usage{
DocumentMetadata(dir.out, title)
}
\arguments{
\item{dir.out}{The directory that you want this file to be saved to.}

\item{title}{Name of the project, to appear at the top of the page.}
}
\description{
A function to document your metadata for people who hate code things but insist they want to know what you are doing
}
\examples{
DocumentMetadata(dir.out = getwd(), title = "Test")

}
\author{
Emily Markowitz Emily.Markowitz@noaa.gov
}
\keyword{documentation}
\keyword{packages}
